fix: resolve 24 review issues across macOS, iOS, and FramerCore
macOS fixes (8):
- ExportBar: URL(string:) → URL(fileURLWithPath:) for paths with spaces
- CanvasView: balance startAccessingSecurityScopedResource with stop
- LayerListSection: undo closures capture binding instead of stale struct
- PreviewViewModel: reset isLoading on Task cancellation
- FilmstripThumbnail: raise non-selected opacity from 0.55 to 0.7
- ContentView: replace .opacity(0) with conditional if for BeforeAfterToggle
- LayerListSection: overlay thumbnail NSImage size in points not pixels
- PresetPreviewGrid: serial preset rendering instead of N concurrent tasks

iOS fixes (10):
- EditorView: only append PhotoItem if data.write succeeds
- EditorView: clear selectedPickerItems on empty selection
- EditorView: defer isLoadingPhotos = false for cancellation safety
- EditorView: move .navigationDestination to NavigationStack owner
- LayerStrip: dynamic index lookup in layer binding (stale index fix)
- PresetPreviewCache: serial rendering with single FrameProcessor
- EditorView: wrap thumbnail loading in Task.detached
- BottomPanel: safe area padding for home indicator
- LayerDetailView: convert Color to sRGB before hex extraction
- PreviewArea: replace infinity long-press with DragGesture

FramerCore fixes (5):
- BorderRenderer: canCoalesce scans full run, not just first 2 layers
- PresetStore: replace force-unwrap with fallback to temp directory
- FrameProcessor: CGFloat division in applyRotation for odd dimensions
- TextureFrameProvider: hold lock across cache check+populate (TOCTOU)
- DitherRenderer: cap pixel scale to 32 on export path

+++ Sources/FramerCore/Processing/FrameProcessor.swift
@@ -133,8 +133,8 @@ public actor FrameProcessor {
                                   bitmapInfo: image.bitmapInfo.rawValue) else { return image }
         ctx.translateBy(x: CGFloat(newW) / 2, y: CGFloat(newH) / 2)
         ctx.rotate(by: -radians)
-        ctx.draw(image, in: CGRect(x: -image.width / 2, y: -image.height / 2,
-                                    width: image.width, height: image.height))
+        ctx.draw(image, in: CGRect(x: -CGFloat(image.width) / 2, y: -CGFloat(image.height) / 2,
+                                    width: CGFloat(image.width), height: CGFloat(image.height)))
         return ctx.makeImage() ?? image
     }
